requirements.txt
asgiref==3.4.1
Django==3.2.9
dnspython==2.1.0
pymongo==4.0
pytz==2021.3
sqlparse==0.4.2
djangorestframework~=3.12.4
Pillow~=8.4.0
google-cloud-secret-manager==2.8.0
django-crispy-forms==1.13.0
django-address==0.2.5